requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.36
psycopg[binary]==3.3.2
jinja2==3.1.3
python-multipart==0.0.9
httpx==0.27.0
pydantic-settings==2.6.1
greenlet==3.1.1
alembic==1.13.1
itsdangerous==2.1.2
